+++ tests/test-0753webm_block_addition_id_4.rb
@@ -5,3 +5,5 @@ describe "mkvmerge / WebM with block addition ID 4, auto-creating mappings for i
 
 test_merge "data/mkv/hdr10_plus_vp9_sample_ffmpeg_output.mkv"
 test_merge "data/webm/hdr10_plus_vp9_sample.webm"
+test_merge "data/mkv/hdr10_plus_vp9_sample_ffmpeg_output.mkv", :args => "--webm"
+test_merge "data/webm/hdr10_plus_vp9_sample.webm",             :args => "--webm"